requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.1.0
xhtml2pdf==0.2.11
django-filter==23.5
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-dotenv==1.0.1
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1